requirements.txt
python-telegram-bot==20.7
scikit-learn
numpy